chore(docs): improve quickstart guide

+++ scripts/vault.sh
@@ -1,5 +1,8 @@
-#!/usr/bin/bash
+#!/usr/bin/env bash
 set -x
+
+command -v vault >/dev/null 2>&1 || { echo "vault is not installed.  Aborting." >&2; exit 1; }
+
 # kill any remaining vault instances
 kill $(pgrep -x vault) || true
 
@@ -15,22 +18,6 @@ export VAULT_TOKEN="root"
 # enable transit engine
 vault secrets enable transit
 vault write -f transit/keys/kms
-
-# create sa, secret and crb
-#kubectl apply -f scripts/rbac.yml
-
-# # enable k8s auth on kubernetes
-# token=$(kubectl get secret -n kube-system vault-auth -o go-template='{{ .data.token }}' | base64 --decode)
-# ca_cert=$(kubectl get cm kube-root-ca.crt -o jsonpath="{['data']['ca\.crt']}")
-
-# # enabel k8s auth on vault
-# vault auth enable kubernetes
-# vault write auth/kubernetes/config \
-#     token_reviewer_jwt="${token}" \
-#     kubernetes_host="https://127.0.0.1:8443" \
-#     kubernetes_ca_cert="${ca_cert}"
-
-# vault write auth/kubernetes/role/kms bound_service_account_names=default bound_service_account_namespaces=kube-system policies=kms ttl=24h
 
 # write vault policy
 vault policy write kms - <<EOF
